requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
yt-dlp==2023.11.16
openai-whisper==20231117
openai==1.3.7
python-multipart==0.0.6 